move lazy image and video to common components folder
